requirements.txt
asgiref==3.11.0
colorama==0.4.6
Django==5.2.4
django-cors-headers==4.9.0
django-filter==25.2
django-graphql-playground==1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.7
graphql-relay==3.2.0
inflection==0.5.1
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
promise==2.3
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
sqlparse==0.5.4
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
